Added section to remove "Interaction" nodes from the edge network
These nodes describe phosphorylations and such but are not really relevant for the connectivity of the network, except if they are the link between multiple other nodes. Therefore, added logic to determine to how many other nodes an "Interaction" node is connected to and then generate new source-target pairs so that all Interaction nodes can be removed. Also specified delimiter for edge list import as TAB