Added SPARQL output automation for WikiPathways
The query to be sent to the endpoint is stored in a local txt file which is read, converted to URL, sent to the endpoint and downloaded as HTML, enabling automatic requests to the endpoint. Must still be implemented for the pathway and edge list queries, and potentially the AOP-Wiki queries.